refactor: material package creation
