+++ material/panel.py
@@ -1,0 +1,30 @@
+import bpy
+from bpy.types import Panel
+
+class NTPMaterialPanel(Panel):
+    bl_label = "Material to Python"
+    bl_idname = "NODE_PT_mat_to_python"
+    bl_space_type = 'NODE_EDITOR'
+    bl_region_type = 'UI'
+    bl_context = ''
+    bl_category = "NodeToPython"
+
+    @classmethod
+    def poll(cls, context):
+        return True
+    
+    def draw_header(self, context):
+        layout = self.layout
+
+    def draw(self, context):
+        layout = self.layout
+        row = layout.row()
+        
+        # Disables menu when there are no materials
+        materials = [mat for mat in bpy.data.materials if mat.node_tree]
+        materials_exist = len(materials) > 0
+        row.enabled = materials_exist
+        
+        row.alignment = 'EXPAND'
+        row.operator_context = 'INVOKE_DEFAULT'
+        row.menu("NODE_MT_ntp_material", text="Materials")
